[#7]fix: amount 자료형 Long으로 변경

+++ src/main/java/org/example/mollyapi/payment/dto/request/PaymentCancelReqDto.java
@@ -3,7 +3,7 @@ package org.example.mollyapi.payment.dto.request;
 public record PaymentCancelReqDto (
         String paymentKey,
         String cancelReason,
-        Integer cancelAmount
+        Long cancelAmount
 //        String refundReceiveAccount
 ){
 }
